# Refactor smoothing for reuse and fewer artefacts