Our polygon smoothing works by expanding and then contracting each shape.

Because it was using different parameters (resolution and join style) for the expansion and contraction it was introducing artefacts, i.e. it was creating entirely new shapes. This seemed to be mostly happening when the original polygons were very small, like those used by the Flood Warning Service.

This commit refactors the smoothing to re-use our existing `bleed_by` method to expand and contract the polygon. By re-using this method we have the same parameters for both the expansion and contraction. This seems to prevent the artefacts.

# `original_polygons`

<img width="1403" alt="Screenshot 2021-08-09 at 10 16 47" src="https://user-images.githubusercontent.com/355079/128684775-c7712a54-99ea-4daf-b577-8c4e680b4521.png">

# `original_polygons.smooth.simplify` (before)

Note the artefacts around the centre of Penzance and St Michael’s Mount

<img width="1401" alt="Screenshot 2021-08-09 at 10 18 37" src="https://user-images.githubusercontent.com/355079/128684789-f52fc76b-07e7-4a5a-9206-ffc68378cc2b.png">

# `original_polygons.smooth.simplify` (after)

<img width="1401" alt="Screenshot 2021-08-09 at 10 19 33" src="https://user-images.githubusercontent.com/355079/128684795-1f38e49a-c668-41f5-8721-f50e7c2bd84c.png">

# Side by side 

Original | Before | After
---|---|---
<img width="1403" alt="Screenshot 2021-08-09 at 10 16 47" src="https://user-images.githubusercontent.com/355079/128684775-c7712a54-99ea-4daf-b577-8c4e680b4521.png"> | <img width="1401" alt="Screenshot 2021-08-09 at 10 18 37" src="https://user-images.githubusercontent.com/355079/128684789-f52fc76b-07e7-4a5a-9206-ffc68378cc2b.png"> | <img width="1401" alt="Screenshot 2021-08-09 at 10 19 33" src="https://user-images.githubusercontent.com/355079/128684795-1f38e49a-c668-41f5-8721-f50e7c2bd84c.png">

# Performance

```shell
python -m timeit -s "from notifications_utils.polygons import Polygons; from tests.test_polygons import LEA_VALLEY" "Polygons([LEA_VALLEY]).smooth"
```

## Before 

1000 loops, best of 3: 189 usec per loop

## After 

1000 loops, best of 3: 210 usec per loop

Small but not significant slowdown, likely caused by `union`ing the polygons twice (once per call to `bleed_by`) rather than once. Here’s what the internals of the `bleed_by` method look like:
https://github.com/alphagov/notifications-utils/blob/83ed2955876de88e3e14a107a48ec4271aba80d9/notifications_utils/polygons.py#L190-L202